+++ src/prefect/config.toml
@@ -139,7 +139,7 @@ checkpointing = false
     [engine.executor]
 
     # the default executor, specified using a full path
-    default_class = "prefect.engine.executors.LocalExecutor"
+    default_class = "prefect.executors.LocalExecutor"
 
         [engine.executor.dask]
         # the default scheduler address for the DaskExecutor.
