+++ src/prefect/engine/__init__.py
@@ -1,6 +1,6 @@
 from warnings import warn
 from prefect import config
-import prefect.engine.executors
+import prefect.executors
 import prefect.engine.state
 import prefect.engine.signals
 import prefect.engine.result
@@ -26,9 +26,9 @@ def get_default_executor_class() -> type:
         except ValueError:
             warn(
                 "Could not import {}; using "
-                "prefect.engine.executors.LocalExecutor instead.".format(config_value)
+                "prefect.executors.LocalExecutor instead.".format(config_value)
             )
-            return prefect.engine.executors.LocalExecutor
+            return prefect.executors.LocalExecutor
     else:
         return config_value
