+++ tests/engine/cloud/test_cloud_flows.py
@@ -9,7 +9,7 @@ import pytest
 import prefect
 from prefect.client.client import Client, FlowRunInfoResult, TaskRunInfoResult
 from prefect.engine.cloud import CloudFlowRunner, CloudTaskRunner
-from prefect.engine.executors import LocalExecutor
+from prefect.executors import LocalExecutor
 from prefect.engine.result_handlers import JSONResultHandler, ResultHandler
 from prefect.engine.results import LocalResult
 from prefect.engine.state import (
