+++ tests/engine/test_flow_runner.py
@@ -13,7 +13,7 @@ import prefect
 from prefect.core import Flow, Parameter, Task
 from prefect.engine import signals
 from prefect.engine.cache_validators import duration_only
-from prefect.engine.executors import Executor, LocalExecutor
+from prefect.executors import Executor, LocalExecutor
 from prefect.engine.flow_runner import ENDRUN, FlowRunner, FlowRunnerInitializeResult
 from prefect.engine.result import NoResult, Result
 from prefect.engine.state import (
