+++ tests/environments/execution/test_k8s_job_environment.py
@@ -9,7 +9,7 @@ import yaml
 
 import prefect
 from prefect import Flow
-from prefect.engine.executors import LocalDaskExecutor
+from prefect.executors import LocalDaskExecutor
 from prefect.environments import KubernetesJobEnvironment
 from prefect.storage import Docker
 from prefect.utilities.configuration import set_temporary_config
@@ -71,7 +71,7 @@ def test_create_k8s_job_environment(job_spec_file):
 
 def test_create_k8s_job_environment_with_deprecated_executor_kwargs(job_spec_file):
     with set_temporary_config(
-        {"engine.executor.default_class": "prefect.engine.executors.LocalDaskExecutor"}
+        {"engine.executor.default_class": "prefect.executors.LocalDaskExecutor"}
     ):
         with pytest.warns(UserWarning, match="executor_kwargs"):
             environment = KubernetesJobEnvironment(
