+++ tests/executors/test_executors.py
@@ -12,12 +12,23 @@ import distributed
 import pytest
 
 import prefect
-from prefect.engine.executors import (
+from prefect.executors import (
     DaskExecutor,
     Executor,
     LocalDaskExecutor,
     LocalExecutor,
 )
+
+
+@pytest.mark.parametrize(
+    "cls_name", ["LocalExecutor", "LocalDaskExecutor", "DaskExecutor"]
+)
+def test_deprecated_executors(cls_name):
+    old_cls = getattr(prefect.engine.executors, cls_name)
+    new_cls = getattr(prefect.executors, cls_name)
+    with pytest.warns(UserWarning, match="has been moved to"):
+        obj = old_cls()
+    assert isinstance(obj, new_cls)
 
 
 class TestBaseExecutor:
